Separate fileManipulation functions
