+++ PapersCited/citationAnalysis.py
@@ -1,20 +1,8 @@
 import locale
 locale.setlocale(locale.LC_ALL, "")
 
-import os
-import sys
-import textract
-import xlsxwriter
-
 # regex in Python
 import regex
-
-# GUI elements - for the file dialog. Explicitly import the filedialog submodule.
-import tkinter
-from tkinter import filedialog
-from tkinter import messagebox
-
-from docx2python import docx2python
 
 # In the actual string, a single \ is used. But for escaping it, we need to put
 # \\ inside strings. Otherwise it will append lines, causing indentation errors.
@@ -220,94 +208,6 @@ class CitationType:
 
 # FUNCTIONS ----
 
-# Open a file dialog to select a file to process.
-def get_file():
-    root = tkinter.Tk()
-    # Make the main window of tkinter invisible since we only need the dialog.
-    root.withdraw()
-    root.attributes("-topmost", True)
-    filename = filedialog.askopenfilename(title = "Select a document to search for citations:")
-    root.destroy()
-    return(filename)
-
-
-def check_file(filename):
-    # Check if the file entered actually exists
-    file_exists = os.path.isfile(filename)
-
-    # If the filename is False, exit the program.
-    if file_exists == False:
-        input("No file selected. Press Enter to end the program.")
-        sys.exit()
-
-    file_extension = os.path.splitext(filename)[1]
-    if file_extension.casefold() == ".pdf":
-        print("Warning!\nReading PDF files is not recommended and might result in inaccurate transcription.\n")
-
-    if file_extension.casefold() == ".txt":
-        print("Warning! Reading .txt files might lead to problems with special characters." +
-              "\nTo ensure the best format is used, backup the .txt file, \
-                then try saving it in UTF-8 or ANSI encoding." +
-              "\n(\"Save as...\" dialog, \"Encoding:\" at the bottom.)\n")
-
-
-def read_docx_footnotes(filename):
-    content =  docx2python(filename)
-    footnotes_container = content.footnotes_runs
-    content.close()
-    
-    if footnotes_container == []:
-        return("")
-    
-    # The third nested list contains different footnotes,
-    # that footnote's nested list [0][1] is the footnote text
-    footnotes = []
-    footnotes_unnested = footnotes_container[0][0]
-    for footnote in footnotes_unnested:
-        footnotes.append(footnote[0])
-    
-    # First two lists are always empty and have no [1] object
-    footnotes = footnotes[2 : (len(footnotes))]
-    
-    footnotes = [footnote[1] for footnote in footnotes]
-    
-    footnotes_text = " \n ".join(footnotes)
-    
-    return(footnotes_text)
-
-def read_document(filename):
-
-    try:
-        target_document = textract.process(filename, output_encoding="utf-8-sig")
-    except Exception as e:
-        # If the file exists, but cannot be read, an error will be raised.
-        print(
-            f"The file {filename} couldn't be read. Make sure the file is a valid textual file.")
-        print("If you can regularly open it, you may be missing certain libraries:")
-        print("antiword for .doc (not .docx)")
-        print("poppler for .pdf")
-        print("\nPlease check 'help_with_libraries.txt' at PapersCited Github:")
-        print("https://github.com/Mkranj/PapersCited/blob/main/help_with_libraries.txt")
-        print("\nThe error message:\n")
-        print(e)
-        input("\nPress Enter to exit the program.")
-        sys.exit()
-
-    # UTF-8 encoding so it recognises foreign characters
-    target_document = target_document.decode("utf-8-sig")
-    
-    file_extension = os.path.splitext(filename)[1]
-    if file_extension.casefold() == ".pdf":
-        target_document = target_document.replace("\r\n", " ")
-        target_document = target_document.replace("\r", "")
-        target_document = target_document.replace("\n", " ")
-    
-    if file_extension.casefold() == ".docx":
-        footnote_text = read_docx_footnotes(filename)
-        target_document = target_document + " \n " + footnote_text
-        
-    return(target_document)
-
 def get_matches_solo_author(text, drop_excluded_phrases = False):
     rx = RegexPatterns()
     matches = regex.findall(
@@ -372,94 +272,7 @@ def get_matches_two_surnames_et_al(text, drop_excluded_phrases = False):
     matches = CitationType(matches)
     if drop_excluded_phrases: matches.drop_excluded_phrases()
     return(matches)
-
-def find_citations(filename):
-    check_file(filename)
-    document = read_document(filename)
     
-    # Get all types of citations
-    solo_authors = get_matches_solo_author(document, drop_excluded_phrases = True)
-    two_authors = get_matches_two_authors(document, drop_excluded_phrases = True)
-    three_authors = get_matches_three_authors(document, drop_excluded_phrases = True)
-    author_et_al = get_matches_author_et_al(document, drop_excluded_phrases = True)
-    two_surnames = get_matches_two_surnames(document, drop_excluded_phrases = True)
-    two_surnames_et_al = get_matches_two_surnames_et_al(document, drop_excluded_phrases = True)
-        
-    solo_authors.delete_clones_of_citations(two_authors)
-    
-    narrower_citations = CitationType(solo_authors.citations + 
-                                      two_authors.citations +
-                                      author_et_al.citations)
-    
-    wider_citations = CitationType(three_authors.citations + 
-                                   two_surnames.citations +
-                                   two_surnames_et_al.citations)
-    
-    narrower_citations.cleanup()
-    wider_citations.cleanup()
-    
-    citation_list = [
-        narrower_citations,
-        wider_citations
-    ]
-    return(citation_list)
-    
-            
-def write_excel(filename, citations, wider_citations):
-    # Retrieve the filepath (and extension) of the analysed document,
-    # The output file will have a similar name and be created 
-    # in the same directory.
-    output_file_prefix = os.path.splitext(filename)
-    output_filename = output_file_prefix[0] + "_citations.xlsx"
-
-    # Create a file
-    try:
-        workbook = xlsxwriter.Workbook(output_filename)
-    except:
-        print(f"Cannot create a file at {output_filename}.")
-        print("Possible permissions issue, can you create files at that folder?")
-        input("\nPress Enter to exit the program.")
-        sys.exit()
-
-    worksheet1 = workbook.add_worksheet()
-
-    # We'll write in the second column.
-    # The first column is useful for marking certain citations as "done" or "double-check"
-    column = 1
-    row = 0
-
-    # Iterate through output list
-    for item in citations.citations:
-        worksheet1.write(row, column, item)
-        row += 1
-
-    # If there are potential wider matches, write them in the sixth column.
-    if wider_citations:
-        column = 5
-        row = 0
-        for item in wider_citations.citations:
-            worksheet1.write(row, column, item)
-            row += 1
-    
-    workbook.close()
-    
-    n_narrower_citations = len(citations.citations) 
-    try:
-        n_wider_citations = len(wider_citations.citations) 
-        total_citations = n_narrower_citations + n_wider_citations
-    except:
-        total_citations = n_narrower_citations
-    
-    print("--------------------")
-    print(f"Success! A file with found citations has been created: {output_filename}.")
-    
-    if n_wider_citations:
-        print(f"{n_narrower_citations} citations have been found, along with" +
-              f" {n_wider_citations} wider citations, displayed to the right.")
-    
-    print(f"A total of {total_citations} different citations have been recorded.")
-
-
 def preview_citations(citations, wider_citations):
     print("\n")
     print("Citations found:")
@@ -468,16 +281,6 @@ def preview_citations(citations, wider_citations):
         print("\n")
         print("Wider citations found:")
         [print(citation) for citation in wider_citations.citations]
-
-def dialog_process_another_file():
-    root = tkinter.Tk()
-    # Make the main window of tkinter invisible since we only need the dialog.
-    root.withdraw()
-    root.attributes("-topmost", True)
-    process_another_file = messagebox.askyesno(title = "Process another file?", message = "Choose another file to extract citations from?\
-                           \nPress 'No' to end the program.")
-    root.destroy()
-    return(process_another_file)
 
 def citations_to_string(narrower_citations, wider_citations):
   citation_string = []
